requirements.txt
playwright~=1.55.0
pytest~=8.4.2
pytest-playwright~=0.7.1
Rx~=3.2.0
allure-pytest~=2.13.5
flake8
python-dateutil~=2.9.0
pytz~=2025.2